Add support for arbitrary guards
This is a proof of concept for supporting arbitrary function calls
as erlang guards. The goal is to combine it with pureness testing,
and only allow functions without side-effects.

The current approach disables certain checks in erl_lint, making
them warnings instead. Support for higher order calls is not yet
implemented. core_lint most probably needs updating too.

Minor additions were required in v3_kernel, but most of the changes
are located in v3_life, where top level match expressions are wrapped
in blocks when such guard expressions are detected. This enables the
creation of the proper allocate_zero instructions in v3_codegen.
Deleting registers which are present in the stack is also required
in v3_codegen, in order to generate move instructions in guard fail
labels.

The patch should not affect code generation if no such expressions
are present. However, because of a change made in beam_validator,
the no_exception_in_catch test case in beam_validator_SUITE now
fails.