Add checks for pure guards
Make user defined guards optional, and only allow them if they are
side-effect free.

Support for such guards is enabled with the `+pure_guards' compiler
option.

This requires an extra compiler pass which analyses the module being
compiled for side-effects. Impure guards are reported as errors, while
pure ones are reported as warnings.

In addition to `+pure_guards', the compiler options `+{plt,"path"}' and
`+impure_exceptions' are now recognised. The first one specifies a
persistent lookup table with pureness results to be loaded (necessary to
bootstrap the analysis with values for built-in functions and bifs),
while the second makes the analysis treat functions which may raise
exceptions as impure (considered pure by default since guards are
wrapped in try/catch).

A copy of the `purity' module is required in order to try out pure
guards.