Automatical generation of 'behaviour_info' function from '-callback' attributes.
'behaviour_info(callbacks)' is a special function that is defined in a module
which describes a behaviour and returns a list of its callbacks.

This function is now automatically generated using the '-callback' specs. An
error is returned by lint if user defines both '-callback' attributes and
the behaviour_info/1 function. If no type info is needed for a callback use
a generic spec for it.