New code generation for user defined guards
The previous approach only worked for a limited number of cases. This
approach covers more, but is still incomplete.

The biggest problem was generating instructions to restore x registers
to their state before the guard call, in the guard fail labels. This
approach generates these instructions at a separate label, replacing the
guard fail label with this one. At the end of the separate label there
is a jump back to the original guard fail label.

There are still issues with complex guard expressions which require
guard_match/guard_break blocks. These cause a compiler crash atm.

Another issue not dealt with regards protected blocks, which remove the
try-catch wrapper around the guard. This causes exceptions in user
defined guards to escape instead of being mapped to false.

Examples of working and broken code are placed in the gt_working and
gt_broken directories respectively.