Fix stack trace handling on PPC64
Function pointers on PPC64 point to a special function descriptor
instead of a function entry point. As a workaround, we sometimes
handle symbols pointing to code as pointers to ints when we need to
compare addresses in native code to addresses in the PPC glue. This
doesn't normally cause any problems, but we have to be careful when we
take the address of such symbols.

Writing (unsigned long)foo instead of (unsigned long)&foo only works
as long as foo is declared as function or an array. This patch changes
the behavior of the AEXTERN macro on PPC64 so that external assembly
symbols are declared as arrays of ints instead of a single int. This
fixes a crash that may occur whenever HiPE tried to create a stack
trace, e.g. when handling some exceptions in native code.