Fix bugs with duplicate opaque declaration
When the declaration of an opaque type is given as an attribute
AND as hardcoded info in erl_*types the result of the
opaque_match_record will contain 2 elements. This breaks the
expected single element list result in almost every case. Usort
is a simple yet probably temporary measure to mitigate this.